Project Cleanup
- cleaned up .gitignore
- removed unnecessary pch.hpp's in subdirectories
- renamed font texture and coordinate files to more appropriate "arial.png" and "arial.csv"
- added config.hpp.in for configuring local file paths to shaders and text textures/coordinates
- Changed top-level CMakeLists to configure "config.hpp" and removed code that copied the shader/text files to the build directory

+++ src/callbacks.cpp
@@ -54,7 +54,7 @@ void scroll_callback(GLFWwindow *window, double xoffset, double yoffset) {
       glfwGetWindowUserPointer(window));
 
   if (!ImGui::IsWindowHovered(ImGuiHoveredFlags_::ImGuiHoveredFlags_AnyWindow))
-    app->Render.Cam.zoom(yoffset);
+    app->Render.Cam.zoom(static_cast<float>(yoffset));
 }
 
 void framebuffer_resize_callback(GLFWwindow *window, int width, int height) {
